official_writeup/tutorial-trivia/game/requirements.txt
flask
pynacl
pytz
uwsgi